move menu initialization into a separate class
